evaluator: eval系の関数を別ファイルに移動する
evaluator.goが肥大化してきたため
